Add plugin module
* Move all code that uses the awscli package into the plugin submodule
* Move error_handler into util and plugin.util
* Rename awscli_login plugin to awscli_login.plugin
